refactor: user 라우트 및 컨트롤러 auth 분리 작업 수행
